Reorganized source files.
