+++ src/lib.rs
@@ -46,19 +46,22 @@
 use std::num::TryFromIntError;
 
 use rand::{CryptoRng, RngCore};
-use subtle::ConditionallySelectable;
+use subtle::{Choice, ConditionallySelectable};
 use thiserror::Error;
 
-pub mod block_value;
 pub mod database;
 pub mod linear_time_oram;
 pub mod path_oram;
+#[cfg(test)]
+mod test_utils;
 pub(crate) mod utils;
 
 pub use crate::path_oram::recursive_secure_path_oram::DefaultPathOram;
 
-#[cfg(test)]
-mod test_utils;
+use rand::{
+    distributions::{Distribution, Standard},
+    Rng,
+};
 
 /// The numeric type used to specify the size of an ORAM block in bytes.
 pub type BlockSize = usize;
@@ -72,6 +75,10 @@ pub trait OramBlock:
     Copy + Clone + std::fmt::Debug + Default + PartialEq + ConditionallySelectable
 {
 }
+
+impl OramBlock for u8 {}
+impl OramBlock for u16 {}
+impl OramBlock for u32 {}
 
 /// A list of error types which are produced during ORAM protocol execution.
 #[derive(Error, Debug)]
@@ -136,5 +143,45 @@ where
         log::debug!("ORAM write: {}", index);
         let callback = |_: &V| new_value;
         self.access(index, callback, rng)
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+#[repr(align(64))]
+/// An `OramBlock` consisting of unstructured bytes.
+pub struct BlockValue<const B: BlockSize>([u8; B]);
+
+impl<const B: BlockSize> BlockValue<B> {
+    /// Instantiates a `BlockValue` from an array of `BLOCK_SIZE` bytes.
+    pub fn new(data: [u8; B]) -> Self {
+        Self(data)
+    }
+}
+
+impl<const B: BlockSize> Default for BlockValue<B> {
+    fn default() -> Self {
+        BlockValue::<B>([0u8; B])
+    }
+}
+
+impl<const B: BlockSize> OramBlock for BlockValue<B> {}
+
+impl<const B: BlockSize> ConditionallySelectable for BlockValue<B> {
+    fn conditional_select(a: &Self, b: &Self, choice: Choice) -> Self {
+        let mut result = BlockValue::default();
+        for i in 0..B {
+            result.0[i] = u8::conditional_select(&a.0[i], &b.0[i], choice);
+        }
+        result
+    }
+}
+
+impl<const B: BlockSize> Distribution<BlockValue<B>> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> BlockValue<B> {
+        let mut result = BlockValue::default();
+        for i in 0..B {
+            result.0[i] = rng.gen();
+        }
+        result
     }
 }
